Use java.util.UUID on JVM
