--- src/jvmMain/kotlin/platform.kt
@@ -1,8 +1,0 @@
-package com.benasher44.uuid
-
-import java.security.SecureRandom
-import kotlin.random.asKotlinRandom
-
-internal actual fun getRandomUuidBytes(): ByteArray = SecureRandom().asKotlinRandom().nextBytes(UUID_BYTES)
-
-internal actual fun <T> T.freeze() = this
